Completed Day 3
NOTE : My solution for day3 part 1 is a little cheaty (as in I did a small part of it manually) but I CBA to 2 out of the 140 lines properly as I already spent 90 mins figuring out how to do the other 138! I know exactly how I would do it but I do not want to waste the effort programming it - I wanted to get part 2 done, lmao!