Completed Day 20
This took me a while not because it was difficult but because it was so complex and finnicky - very fun one though!
Part 2 harder than I thought it would be, but it was basically Day 8p2 again... get the lcm to avoid brute force iteration. I had this idea but didn't think it would work for some reason, thanks to this writeup
https://github.com/mrphlip/aoc/blob/master/2023/20.md
for confirming I was in fact right.