Completed Day 23
Tricky today - my part 1 solution is very optimised but mostly because I just read a lot of stuff on the Longest Path problem then realised the graph is as Directed Acyclic Graph, so implemented the fast algorithm.

Part 2 didn't take long at all - until I realised how long it would take. The compression system was based off an idea from this
https://www.reddit.com/r/adventofcode/comments/18oy4pc/comment/kekhj7f/?context=1
but it still took me a while to implement.